requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.txt
#
aniso8601==7.0.0          # via graphene
asgiref==3.2.7            # via django
django-fileprovider==0.1.4  # via -r requirements.txt
django-filter==2.2.0      # via -r requirements.txt
django-graphene-permissions==0.0.4  # via -r requirements.txt
django-graphql-jwt==0.3.1  # via -r requirements.txt
django==3.0.6             # via -r requirements.txt, django-filter, django-graphql-jwt, graphene-django
graphene-django==2.9.1    # via -r requirements.txt, django-graphql-jwt
graphene==2.1.8           # via graphene-django
graphql-core==2.3.1       # via django-graphql-jwt, graphene, graphene-django, graphql-relay
graphql-relay==2.0.1      # via graphene
promise==2.3              # via graphene-django, graphql-core, graphql-relay
psycopg2-binary==2.8.5    # via -r requirements.txt
pyjwt==1.7.1              # via django-graphql-jwt
pytz==2020.1              # via django
rx==1.6.1                 # via graphql-core
singledispatch==3.4.0.3   # via graphene-django
six==1.14.0               # via django-fileprovider, graphene, graphene-django, graphql-core, graphql-relay, promise, singledispatch
sqlparse==0.3.1           # via django
djangorestframework